refactor: restructure kafka queue
